requirements.txt
beautifulsoup4==4.6.0
bleach==2.1.1
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.10.8
django-braces==1.11.0
django-contrib-comments==1.8.0
django-filter==1.1.0
django-modeltranslation==0.12.1
django-oauth-toolkit==0.12.0
django-rest-swagger==2.1.2
djangorestframework==3.7.1
filebrowser-safe==0.4.7
future==0.16.0
grappelli-safe==0.4.7
gunicorn==19.7.1
html5lib==1.0b10
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
Mezzanine==4.2.3
mezzanine-api==0.6.1
oauthlib==2.0.1
olefile==0.44
openapi-codec==1.3.2
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.11.0
tzlocal==1.4
uritemplate==3.0.0
urllib3==1.22
webencodings==0.5.1
django-webpack-loader==0.2.4

